Added TerminalService class to handle interactions and commands in the Terminal.               - Added methods to copy text to clipboard and execute Git commands.               - Refactored GitChangesSummarizer to utilize TerminalService for clipboard operations.               - Integrated TerminalService with GitService and TerminalGui for command execution and interactions.
